Store the build artifact
The build output is stored in the GitHub as a workflow artifact to be
used as a release candidate.
It can be downloaded after the succesful build and used for publishing.

NOTE: the `build` job now runs only for the main branch - this is to
ensure that only verified, checked and reviewed commits will be included
in the build.