Send general message for a button click
Once any button is clicked on the popup, the MainPopup script will send
the general "execute" message.
This message can have various option of what exactly should be executed,
but this is added only as an argument of general "execute" message.

This message is then caught by the content script, which will execute
the requred actions on the browser sessions.