Move all the app-related source to app/ directory
To ensure that application implementation sources will be separated from
build artifacts or build-related code, or whole package configuration
file, the application implementation has been moved to app/ directory.
This solution is to make the architecture of the package clean and easy
to manage.

NOTE: This move also required the CI workflow to be modified and path
for wbe-ext and linter to be adjusted to the app/ directory of
verification and build.