Add compilation to the CI workflow
To release the fully working package with the ES6 standard coding it is
required to apply the compilation and transpilation to the workflow.

This is done as additional step in which sources are compiled and cached
for production-ready to be packaged.

The reason for doing this for both branch and main is to ensure that
during development sources will be able to be compiled succesfully.