Implement logic for Popup
The MainPopup's logic is implemented in a separate source named
MainPopupLogic which contains all the functions related to the mechanism
of the MainPopup.
The idea is that it can be imported in both MainPopup and tests to be
used. Also such approach allows to reduce the amount of code in the
MainPopup script to only the 'browser' and 'document' operations.