Wrap input and buttons into form
To make the session name input possible to be pulled from the content it
is needed to make it a form. Also it's required to make sure the submit
has been done on the form so the value of input is ready to be read.

Popup content has been refactored to perform as form instead of loosely
connected divs.
Input is left as it was, but all the buttons are made of <button>
instead of <div> with button class.

NOTES:
* it was also required to adapt the css so the button style will be
  available for <button>
* Layout and design of popup is left intact