Content - plugin engine & command parsing (#17)
* Add required permissions

* Send general message for a button click

Once any button is clicked on the popup, the MainPopup script will send
the general "execute" message.
This message can have various option of what exactly should be executed,
but this is added only as an argument of general "execute" message.

This message is then caught by the content script, which will execute
the requred actions on the browser sessions.

* Use className for click target recognition

* Use target's ID as indicator for command message

The command message (now generated by the sendMessageToEngineListener)
is now generated according to the ID of a target of the event comming
from the popup.

* Implement logic for Popup

The MainPopup's logic is implemented in a separate source named
MainPopupLogic which contains all the functions related to the mechanism
of the MainPopup.
The idea is that it can be imported in both MainPopup and tests to be
used. Also such approach allows to reduce the amount of code in the
MainPopup script to only the 'browser' and 'document' operations.

* Correct ESLint and tests

ESLint complained about undefined error so it was corrected by calling
the reportExecutiveScriptError within a lambda function taking `error`
as a built-in variable.

Tests were provided with a mocked version of built-in variables, but
using function and objects declaration instead of `jest.fn`.

* Apply styling after rebase

* Ignore build artiffacts also in ESLint

* Wrap input and buttons into form

To make the session name input possible to be pulled from the content it
is needed to make it a form. Also it's required to make sure the submit
has been done on the form so the value of input is ready to be read.

Popup content has been refactored to perform as form instead of loosely
connected divs.
Input is left as it was, but all the buttons are made of <button>
instead of <div> with button class.

NOTES:
* it was also required to adapt the css so the button style will be
  available for <button>
* Layout and design of popup is left intact

* Read input value for session name and put in command

A name of the session written in input form is read in the popup logic
and placed within the command sent to content engine.

Unit tests check the correction of the command (syntax and parameters).

* Use Content class for content script engine

The engine of the plugin will be implemented by the Content class.
This commit implements the Content with its constructor and methods
returning the basic properties: `command` and `session`.

Constructor is covered with tests.