Do not launch the content script when starting Popup
The content script listens to message sent from the Popup and there's
also the Background engine which also listens to those.
Removing the content script will allow the Background to receive all
runtime messages and respond to them correctly.

The second reason for removing the content script launch from MainPopup
is the Bakcground engine being a future replacement which should
simplify the data flow and component communication in the plugin.