Create the initial database when plugin loads
The `window.onload` opens the database with the initial version 1 and
performs the setup if it is not yet created in the storage.
Global `window` is used so that the created db will be common for all
tabs and will be persisted for a launching device.

The created schema is:
```
{
  sessionName: VARCHAR NOT UNIQUE,
  sessionObject: VARCHAR NOT UNIQUE,
  id: INT NOT NULL,
  PRIMARY KEY (id),
}
```