Open Database with specified db version number
The Database is opened with the number given as an argument so that db
schema is selected according to its version.
The `getSchema` method keeps the array of schemas defined in the system
and returns one schema requested by a caller with certain ID, previously
converted with the `dbNumberToSchemaIndex` method.