Save entry in database with save method
The `save` method is used to write a SINGLE entry in the database:
{ sessionName, sessionData }
which stands for only one certain tab to be written.
This method also checks if it argument matches the database schema and
does nothing if not.

The usage of this method is to inject it to the engine `saveSession`
method which should call this method each time an entry is extracted
from the browser window session data.

NOTE: This commit also adds the `this.instance` property to the Database
class which contains the handler for the opened database.
This is done for the save method to be able to save an entry in the
same, already opened and existing database.