Implement the loading tabs from localStorage
The loading procedure is very simple: take the key of a session - it's
name - and read the keyValue of it from a localStorage.
If it exists, the  Array[] of tabs objects will be returned from this
method,
It it doesn't, the exception will be caught and logged to console, and
undefined will be returned, so no tabs will be reopened.