Place all layout implementation in `<MainPopup/>`
The `<MainPopup/>` element is responsible to render the root of the
popup layout. It currently holds all the legacy HTML implementation.
This has been tested the manually by building it and performing the
"SAVE" action.

NOTE: Both scripts loading React were defaulted to the production mode
so that each time the `install` script is launched, the production-ready
build will be prepared. It does not affect the development speed, but
makes the deployment easier.