Run web-ext verification on build/ instead of app/
The web-ext linter will fail when spotting the JSX syntax.
This syntax is only readable to web-ext linter after transpiling it with
the Babel preset for React. However, this transpiled file is only
available after the compilation of the application, so it is placed in
build/ directory.
Due to that it is needed to run the web-ext linter on the compilation
artiffacts instead of app code.