Compile app sources as a pre-step for web-ext linter
To make a web-ext fully aware of the JSX syntax it's necessary to
compile the application first and run this linter on compiled sources.
But the issue was that the web-ext linter
and compilation steps are running in parallel - making the compiled
sources unavailable for linter.
This commit makes the compilation pre-requisite for the web-ext linter
step.