Migrate the application code to React (#29)
* Install React compiler - Babel

* Configure Babel to consider also react preset

* Update Babel/core and web-ext

* Install react and react-dom

React is installed at the latest version so that all it's functionality
can be used with JSX syntax.
react-dom is a package that provided the project with a standalone
`render` method to put the initial `<MainPopup/>` element into the root
of HTML file.

* Place all layout implementation in `<MainPopup/>`

The `<MainPopup/>` element is responsible to render the root of the
popup layout. It currently holds all the legacy HTML implementation.
This has been tested the manually by building it and performing the
"SAVE" action.

NOTE: Both scripts loading React were defaulted to the production mode
so that each time the `install` script is launched, the production-ready
build will be prepared. It does not affect the development speed, but
makes the deployment easier.

* Use eslint-plugin-react for JSX syntax

* Simplify the main component code

* Remove remote scripts from the MainPopup HTML

* Run web-ext verification on build/ instead of app/

The web-ext linter will fail when spotting the JSX syntax.
This syntax is only readable to web-ext linter after transpiling it with
the Babel preset for React. However, this transpiled file is only
available after the compilation of the application, so it is placed in
build/ directory.
Due to that it is needed to run the web-ext linter on the compilation
artiffacts instead of app code.

* Use extensionName state for header text

* Fix ESLint errors

* Compile app sources as a pre-step for web-ext linter

To make a web-ext fully aware of the JSX syntax it's necessary to
compile the application first and run this linter on compiled sources.
But the issue was that the web-ext linter
and compilation steps are running in parallel - making the compiled
sources unavailable for linter.
This commit makes the compilation pre-requisite for the web-ext linter
step.

* Move dependencies to devDependencies