Rename `install` script to `deploy`
Having the production ready build script marked as `install` makes it less readable
and doesn't describe the purpose of it. Also, this name causes the build to be
performed each time the *node_modules* are installed, which is undesired.
The `deploy` name says exactly what this script does - it creates the production
ready build.