Create each tab during reopening
When session is reopened with given name, the whole session tab array
object is parsed into separate tab.
This tab has the properties which are used to `browser.tabs.create()` by
giving the exact object containing all required properties.

NOTE: It turned out, that the object to create the tab with should
contain different properties, then the one which the Engine was designed
for.
See:
developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/tabs/create#parameters
Due to that, the `mdnTabToDatabaseTabObject` was redesigned to directly
convert the tab into stored object.
This solution also simplifies and reduces the amount of data in storage.