Fix: undefined session names array when empty
When there's no session name saved in a storage yet, and a component
fetches those data, the empty array should be returned which would
explicitly indicate that there's no session registered yet.
This fixes the situation, where no data in session storage resulted in
`null` or `undefined` instead of array which resulted in crash or
component unable to render.

NOTE:
This fix should be on the main branch along with commit 7f99b29 but it
was discovered here and is required on this branch, so this is a hotfix
to be delivered along with redesign feature implementation.