Use class name for input value instead of name
When providing input component with `name` or `id` all values previously
entered to this input were kept in history resulting in their text
appear as session names suggestions no matter if they were in database
or not.
This could lead to confusion for a user to seeing a name of session
available which he/she already deleted.

To avoid that the `name` and `id` attributes of session name input were
removed, which cause the document querying impossible with these. So to
query for session name value the element with "session-name-input" is
used and it's value is fetched.

NOTE: `getElementsByClassName` returns all elements that uses this
class, so [0] index is used as there's only the input that uses this.