Use React to redesign the popup (#32)
* Create initial (empty) components for popup

* Fix: undefined session names array when empty

When there's no session name saved in a storage yet, and a component
fetches those data, the empty array should be returned which would
explicitly indicate that there's no session registered yet.
This fixes the situation, where no data in session storage resulted in
`null` or `undefined` instead of array which resulted in crash or
component unable to render.

NOTE:
This fix should be on the main branch along with commit 7f99b29 but it
was discovered here and is required on this branch, so this is a hotfix
to be delivered along with redesign feature implementation.

* Use ExpandedSessionListInput as input component

* Style and implement input component with env style

* Use class name for input value instead of name

When providing input component with `name` or `id` all values previously
entered to this input were kept in history resulting in their text
appear as session names suggestions no matter if they were in database
or not.
This could lead to confusion for a user to seeing a name of session
available which he/she already deleted.

To avoid that the `name` and `id` attributes of session name input were
removed, which cause the document querying impossible with these. So to
query for session name value the element with "session-name-input" is
used and it's value is fetched.

NOTE: `getElementsByClassName` returns all elements that uses this
class, so [0] index is used as there's only the input that uses this.

* Fix ESLint errors for components

* Implement ActionButton to use props

* Replace <button/> with ActionButton component

* Disable prop-types rule for React components

* Implement the popup layout

NOTE: Buttons having the icon should consume the click event on their
whole space. The icon then should have the same `id` set so that the
`event.target.id` will match no matter if click is on icon or button
empty space.

* Add text to the button to improve readability

* Move panel-session-name div to ExpandedSessionListInput

* Utilize <Logo/> component with Session Storage name